pcf-calculator/my_function/main.py
import azure.functions as func
from my_function.backend import app 


def main(req: func.HttpRequest, context: func.Context) -> func.HttpResponse:
    return func.WsgiMiddleware(app.wsgi_app).handle(req, context)