pcf-calculator/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
flask
azure-functions
matplotlib
shapely
redis==5.0.0
my_function/costum_package/l26_geometry_reader-0.2-py3-none-any.whl
requests
python-dotenv